Move Array module to Internal
